Change timestamp (#42)
* extended explanation for sessionTimeout
execution_time is now Double
added extra info text to the default session value
bumped version number

* change version number for CocoaPods and Carthage

+++ ExponeaSDK/ExponeaSDK/Classes/Models/Parameters/CustomerParameters.swift
@@ -140,7 +140,7 @@ extension CustomerParameters: RequestParametersType {
             preparedParam["filter"] = .dictionary(filterList)
             
             if let time = data.executionTime {
-                preparedParam["execution_time"] = .int(time)
+                preparedParam["execution_time"] = .double(Double(time))
             }
             
             if let timezone = data.timezone {
